Keep configuration filterControlPanelsSchema
