+++ docs/source/configuration/settings-reference.md
@@ -205,8 +205,11 @@ controlpanels
 
     The group can be one of the default groups 'General', 'Content', 'Security', 'Add-on Configuration', 'Users and Groups' or a custom group.
 
+filterControlPanelsSchema
+    A schema factory for a control panel. It is used internally, to tweak the schemas provided by the controlpanel endpoint, to make them fit for Volto. Uses the ``unwantedControlPanelsFields`` setting.
+
 unwantedControlPanelsFields
-    Control panels fields that are not used on the Volto. It is used internally, to tweak the schemas provided by the controlpanel endpoint, to make them fit for Volto.
+    Control panels fields that are not used on the Volto. It is used internally by ``filterControlPanelsSchema`` function.
 
 errorHandlers
     A list of error handlers that will be called when there is an unhandled exception. Each error handler is a function that
