+++ packages/volto/news/4819.feature
@@ -1,1 +1,1 @@
-Replaces ``filterControlPanelsSchema`` setting by ``unwantedControlPanelsFields`. @wesleybl
+Add setting ``unwantedControlPanelsFields`` and use it in the function ``filterControlPanelsSchema``. @wesleybl
