+++ packages/volto/src/config/index.js
@@ -24,6 +24,7 @@ import { styleClassNameConverters, styleClassNameExtenders } from './Style';
 import {
   controlPanelsIcons,
   filterControlPanels,
+  filterControlPanelsSchema,
   unwantedControlPanelsFields,
 } from './ControlPanels';
 
@@ -152,6 +153,7 @@ let config = {
     controlpanels: [],
     controlPanelsIcons,
     filterControlPanels,
+    filterControlPanelsSchema,
     unwantedControlPanelsFields,
     externalRoutes: [
       // URL to be considered as external
